backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
litellm
reportlab
openpyxl
supabase
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
